Fixed bug of long callbacks timing out filter notifications
If we do not call getFilterChanges in Parity for 60 seconds, it forgets
the filter and subsequent calls will report no new changes. This used to
happen when a `on_block` callback took more than that time.

This is a quick&simple workaround which runs the callback in a separate
thread.