Unit-tested failed token transfers
The main purpose of these tests is to check whether no `Receipt` is
returned by the `transact` and `transact_async` methods when the actual
contract method invocations fail (due to insufficient balance).

We made an assumption some time ago that no `Receipt` will be returned
in such cases. It may turn out this assumption is wrong/misleading, but
for now it is how it is and I want to properly unit-test it as I will be
making changes related to `Receipt` parsing in a minute.