Got rid of explicit nonce management in keepers
The approach of the keeper excplicitely managing its nonce sequence
didn't work very well if transactions from the same account were sent
manually when the keeper was running. It did cause nonce reuse.

In the current approach we capture the nonce the moment after we have
sent the transaction. This way it's the Parity (geth?) node which
assigns nonces, but at the same time we know what nonce was used so we
can easily override pending transactions with higher gas price for
example.