Transaction receipt detectionn changed so it detects competing transaction
In case another bot running on the same address manages to send a
transaction with the same nonce, the keeper used to get stuck
permanently. This change fixes it i.e. if it sees that its nonce number
has been used by somebody else, it just considers it as a failed one.