`get_orders_by_maker()` method added to the Oasis API
This function returns all active orders filtered by maker. Thanks to the
fact that if an order was created by some account (`maker`) it will stay
owned by it forever, this function creates an internal cache of orders
which *do* *not* belong to a specified maker, which allows it to query
less orders and thus return the result faster.