Replaced `threading.Condition` with `threading.Event`
In this new approach id `trigger_event` is being called while the
previous event callback is still operating, we can be sure that the
callback will be run again when it finishes. This is thanks to the
`clear()` function we can call on `threading.Event` before we start the
callback.